pyproject.toml
[project]
name = "ultimate-chinese"
version = "0.1.0"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "brain-brew>=0.3.11",
    "pandas>=2.2.3",
    "py-pinyin-split>=1.2.0",
    "tqdm>=4.67.0",
    "pypinyin>=0.53.0",
]

[dependency-groups]
dev = [
    "ruff>=0.7.3",
    "pyperclip>=1.9.0",
    "azure-cognitiveservices-speech>=1.41.1",
    "python-dotenv>=1.0.1",
    "mypy>=1.13.0",
    "pandas-stubs>=2.2.3.241126",
]

[tool.ruff.lint]
select = [
    # pycodestyle
    "E",
    # Pyflakes
    "F",
    # pyupgrade
    "UP",
    # flake8-bugbear
    "B",
    # flake8-simplify
    "SIM",
    # isort
    "I",
]
